% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cifti.R
\name{simulate_cifti}
\alias{simulate_cifti}
\title{Simulate cifti data for task fMRI}
\usage{
simulate_cifti(
  wb_path,
  ntasks = 2,
  ntime = 300,
  resamp_res = NULL,
  max_amplitude = 1,
  onsets = NULL,
  durations = NULL,
  TR = 1,
  ar_error = NULL,
  surfL = NULL,
  surfR = NULL
)
}
\arguments{
\item{wb_path}{Path to the connectome workbench (required)}

\item{ntasks}{The number of tasks with which to simulate response time series}

\item{ntime}{The length of the response time series}

\item{resamp_res}{The resolution at which to resample the simuluated cifti
data (default \code{resamp_res = 5000}).}

\item{max_amplitude}{The maximum amplitude of the coefficients in percent
signal change (default \code{max_amplitude = 1})}

\item{onsets}{A list of times for the onsets of tasks, each element
representing a task. Overrides \code{ntasks} for determining the number of
tasks if both are given.}

\item{durations}{A list of durations for each task, each element representing
a task. The list should have the same length as \code{onsets}, and each
list element should have the same length as the respective element within
\code{onsets}.}

\item{TR}{The repetition time, in seconds. Must be a whole number.}

\item{ar_error}{A vector of length \code{p} of the AR(p) coefficients for the
error term. The default is \code{NULL}, which results in no autoregressive
error. These coefficient values will be simulated to vary across the
spatial field in order to more accurately reflect most cifti data}

\item{surfL}{Left surface to use for the simulated cifti. If none are given, the
default surfaces from \code{ciftiTools} are used.}

\item{surfR}{Right surface to use for the simulated cifti. If none are given, the
default surfaces from \code{ciftiTools} are used.}
}
\value{
A list of objects containing the simulated data and the relevant
information used to create the data. The simulated response data are found
within the \code{cifti} object. The simulated amplitude fields are found
as cifti objects within the \code{coefficients} object. The simulated
design matrix is found in the \code{design} object.
}
\description{
Simulate cifti data for task fMRI
}
\examples{
\dontrun{
simulate_cifti()
}
}
