% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cifti.R
\name{spatial_effects_cifti}
\alias{spatial_effects_cifti}
\title{Generate spatial effects over areas for a cifti}
\usage{
spatial_effects_cifti(cifti_obj, centers_lambda, smooth_FWHM, max_amplitude)
}
\arguments{
\item{cifti_obj}{a \code{xifti} object}

\item{centers_lambda}{the parameter controlling how many centers the activations will have}

\item{smooth_FWHM}{The full-width half-maximum smoothing value, in mm}

\item{max_amplitude}{The maximum value taken by the spatial effect}
}
\value{
a \code{xifti} object
}
\description{
Generate spatial effects over areas for a cifti
}
\keyword{internal}
