% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cifti.R
\name{spatial_effects_cifti_cs}
\alias{spatial_effects_cifti_cs}
\title{Generate spatial effects over areas for a cifti with cortical surface data}
\usage{
spatial_effects_cifti_cs(
  cifti_obj,
  centers_lambda,
  smooth_FWHM,
  max_amplitude,
  n_tasks,
  n_subjects = 1,
  n_sessions = 1,
  n_runs = 1,
  subjects_var = 4,
  sessions_var = 2,
  runs_var = 1
)
}
\arguments{
\item{cifti_obj}{a \code{xifti} object}

\item{centers_lambda}{the parameter controlling how many centers the activations will have}

\item{smooth_FWHM}{The full-width half-maximum smoothing value, in mm}

\item{max_amplitude}{The maximum value taken by the spatial effect}

\item{n_tasks}{The number of tasks to generate}

\item{n_subjects}{The number of subjects for which to generate data}

\item{n_sessions}{The number of sessions per subject to generate}

\item{n_runs}{The number of runs per session to generate}

\item{subjects_var}{The amount of average variance in coefficient location
from subject to subject}

\item{sessions_var}{The amount of average variance in coefficient location
from session to session within subject}

\item{runs_var}{The amount of average variance in coefficient location from
run to run within session}
}
\value{
a list of \code{xifti} objects nested by subject, session, and run
}
\description{
Generate spatial effects over areas for a cifti with cortical surface data
}
\keyword{internal}
